refactor training script | load datasets with AggregatedDataset class
